build: windows 支持
支持 windows 编译，现只支持release版本

+++ src/plugin.h
@@ -18,8 +18,8 @@ public:
 public:
     int  plugin_init();
     void plugin_connection_start(connection_data* key);
-    void plugin_connection_send(connection_data* key, const char* buffer, int len);
-    void plugin_connection_recv(connection_data* key, const char* buffer, int len);
+    void plugin_connection_send(connection_data* key, const char* buffer, size_t len);
+    void plugin_connection_recv(connection_data* key, const char* buffer, size_t len);
     void plugin_connection_close(connection_data* key);
     void plugin_close();
